+++ src/calmjs/tests/test_runtime.py
@@ -363,6 +363,11 @@ class ToolchainRuntimeTestCase(unittest.TestCase):
     Test cases for the toolchain runtime.
     """
 
+    def setUp(self):
+        remember_cwd(self)
+        self.cwd = mkdtemp(self)
+        os.chdir(self.cwd)
+
     def test_toolchain_runtime_basic_config(self):
         tc = toolchain.NullToolchain()
         rt = runtime.ToolchainRuntime(tc)
@@ -374,6 +379,7 @@ class ToolchainRuntimeTestCase(unittest.TestCase):
         self.assertTrue(isinstance(rt.create_spec(), toolchain.Spec))
 
     def test_standard_run(self):
+        stub_stdouts(self)
         tc = toolchain.NullToolchain()
         rt = runtime.ToolchainRuntime(tc)
 
@@ -483,6 +489,77 @@ class ToolchainRuntimeTestCase(unittest.TestCase):
         self.assertTrue(isinstance(result, toolchain.Spec))
         # prove that it did at least run
         self.assertIn('build_dir', result)
+        self.assertEqual(result['link'], 'linked')
+
+    def test_prompted_execution_with_working_dir_flag(self):
+        stub_check_interactive(self, True)
+        stub_stdouts(self)
+        stub_stdin(self, u'y\n')
+
+        target_dir = mkdtemp(self)
+        export_target = join(target_dir, 'export_file')
+        open(export_target, 'w').close()  # write an empty file
+
+        tc = toolchain.NullToolchain()
+        rt = runtime.ToolchainRuntime(tc)
+        result = rt([
+            '--export-target', 'export_file', '--working-dir', target_dir])
+        self.assertTrue(isinstance(result, toolchain.Spec))
+
+        self.assertIn(
+            "export target '%s' already exists, overwrite? " % export_target,
+            sys.stdout.getvalue()
+        )
+
+        # also check the logs.
+        err = sys.stderr.getvalue()
+        self.assertNotIn('CRITICAL', err)
+        self.assertIn('WARNING', err)
+        self.assertIn(
+            "WARNING calmjs.toolchain realpath of 'export_target' resolved to",
+            err)
+        self.assertIn(export_target, err)
+        # prove that it did at least run
+        self.assertEqual(result['export_target'], export_target)
+        self.assertEqual(result['link'], 'linked')
+
+    def test_prompted_execution_without_export_target(self):
+        class ToolchainRuntime(runtime.ToolchainRuntime):
+            def create_spec(self, **kwargs):
+                if not kwargs.get('export_target'):
+                    kwargs['export_target'] = join(
+                        kwargs.get('working_dir') or self.cwd,
+                        'default_location'
+                    )
+                return super(ToolchainRuntime, self).create_spec(**kwargs)
+
+        stub_check_interactive(self, True)
+        stub_stdouts(self)
+        stub_stdin(self, u'y\n')
+
+        target_dir = mkdtemp(self)
+        # write an empty file at expected location
+        export_target = join(target_dir, 'default_location')
+        open(export_target, 'w').close()
+
+        tc = toolchain.NullToolchain()
+        rt = ToolchainRuntime(tc)
+        result = rt(['--working-dir', target_dir])
+        self.assertIn(
+            "export target '%s' already exists, overwrite? " % export_target,
+            sys.stdout.getvalue()
+        )
+        # also check the logs.
+        err = sys.stderr.getvalue()
+        self.assertNotIn('CRITICAL', err)
+        self.assertNotIn('WARNING', err)
+        # since it was joined with a proper working dir.
+        self.assertNotIn(
+            "WARNING calmjs.toolchain realpath of 'export_target' resolved to",
+            err)
+        self.assertNotIn(export_target, err)
+        # prove that it did at least run
+        self.assertEqual(result['export_target'], export_target)
         self.assertEqual(result['link'], 'linked')
 
     def test_prompted_execution_exists_overwrite(self):
@@ -627,18 +704,29 @@ class ToolchainRuntimeTestCase(unittest.TestCase):
         self.assertIn("spec missing key 'export_target';", stderr)
 
     def test_spec_nodebug(self):
+        stub_stdouts(self)
         tc = toolchain.NullToolchain()
         rt = runtime.ToolchainRuntime(tc)
-        result = rt(['--export-target', 'dummy'])
+        with pretty_logging(logger='calmjs', stream=mocks.StringIO()) as s:
+            result = rt(['--export-target', 'dummy'])
         self.assertEqual(result['debug'], 0)
+        self.assertIn(
+            "'export_target' resolved to '%s'" % join(self.cwd, 'dummy'),
+            s.getvalue())
 
     def test_spec_debugged(self):
         tc = toolchain.NullToolchain()
         rt = runtime.ToolchainRuntime(tc)
-        result = rt(['-dd', '--export-target', 'dummy'])
+        with pretty_logging(logger='calmjs', stream=mocks.StringIO()) as s:
+            result = rt(['-dd', '--export-target', join(self.cwd, 'dummy')])
         self.assertEqual(result['debug'], 2)
+        # also test for the export_target resolution
+        self.assertNotIn(
+            "'export_target' resolved to '%s'" % join(self.cwd, 'dummy'),
+            s.getvalue())
 
     def test_spec_debugged_via_cmdline(self):
+        stub_stdouts(self)
         stub_item_attr_value(
             self, mocks, 'dummy',
             runtime.ToolchainRuntime(toolchain.NullToolchain()),
@@ -649,8 +737,13 @@ class ToolchainRuntimeTestCase(unittest.TestCase):
             ],
         })
         rt = runtime.Runtime(working_set=working_set, prog='calmjs')
-        result = rt(['tool', '--export-target', 'dummy', '-d'])
+        with pretty_logging(logger='calmjs', stream=mocks.StringIO()) as s:
+            result = rt(['tool', '--export-target', 'dummy', '-d'])
         self.assertEqual(result['debug'], 1)
+        # also test for the export_target resolution
+        self.assertIn(
+            "'export_target' resolved to '%s'" % join(self.cwd, 'dummy'),
+            s.getvalue())
 
     def test_spec_optional_advice(self):
         from calmjs.registry import _inst as root_registry
@@ -691,7 +784,8 @@ class ToolchainRuntimeTestCase(unittest.TestCase):
         self.assertIn('failure encountered while setting up advices', err)
 
         # Doing it normally should not result in that optional key.
-        result = rt(['--export-target', 'dummy'])
+        with pretty_logging(logger='calmjs', stream=mocks.StringIO()):
+            result = rt(['--export-target', 'dummy'])
         self.assertNotIn('dummy', result)
 
     def test_spec_advise_debugger(self):
@@ -797,7 +891,7 @@ class ToolchainRuntimeTestCase(unittest.TestCase):
         rt = runtime.ToolchainRuntime(tc)
 
         result = rt([
-            '--export-target', 'dummy',
+            '--export-target', join(mkdtemp(self), 'dummy'),
             '--optional-advice', 'example.package',
         ])